refactor(shared/graph_handlers): 단일 파일을 sub-package 로 분해 (#23)
서브모듈 구성:

  config.py          env 기반 자원 관리 튜닝 (timeout / keepalive)
  session.py         ChatContext + log_session (lifecycle 로깅)
  parse.py           요청 파싱 / LLM 응답 텍스트 추출
  factories.py       A2A Task / 이벤트 모델 조립 + 에러 텍스트
  envelope.py        JSON-RPC / SSE envelope 직렬화
  stream.py          graph.astream → SSE 라인 번역 (S1 polling · S2 keepalive)
  send_message.py    GraphSendMessageHandler (단방향)
  send_streaming.py  GraphSendStreamingMessageHandler (SSE)
  __init__.py        두 Handler 클래스만 외부 노출

각 파일 35~150줄. 외부 import 경로 (`dev_team_shared.a2a.server.graph_handlers`)
는 그대로라 consumer (Primary `server.py`) 변경 없음.

logger 이름은 모듈 경로를 따라 `...graph_handlers.session` 등으로 세분화 →
sse_session.start/end 로그가 어느 서브모듈에서 났는지 한눈에 보임.

검증: AgentCard / SendMessage / SendStreamingMessage 정상. lifecycle 로그에
completed · client_disconnect · graph_error 모두 정상 출력. SSE 이벤트 순서
(initial Task → artifact-update × N → status-update final) 보존.

문서:
  - shared/a2a/server/README.md — 서브모듈 표 추가
  - docs/sse-connection.md §6 — S1~S4 매핑을 새 파일 경로로 갱신
